HttpRequestSenderAndReceiverTest/requirements.txt
flask==2.0.2
requests==2.26.0
pillow==8.3.2
numpy==1.21.2
